chore: move annotations tests to correct files
